requirements.txt
flask==3.1.0
joblib==1.4.2
pandas==2.2.3
python-dotenv==1.1.0
requests==2.32.3
scikit-learn==1.6.1
